New textures, new prefabs, new particles, oh my!
We've got a ned station prefab, a new penguin prefab, and a new ushcart prefab!

-Station is completely new, has no animations
-The updated pushcart has benn re-mapped for UVs and has a texture applied
-The updated penguin has an additional animation for jumping off the train

-The old versions of the FBXs have been moved to a new folder just for them, to avoid confusion with naming! (I didn't want to straight-up delete them in case it broke things)
The "Old versions of FBXs" folder can be found in the general FBX resources folder.

The prefabs didn't have as much of a porblem with naming, so the new prefabs are in the usual folder!

What else....
Oh! New particles!
We've got a "spalsh" effect for penguins that get smacked by a fish (or whatever else you need a splash for) - It's 20 particles at max, so feel free to really spam it!

The other one is a new multi-system particle effect! It's for smoke coming from the engine. It includes:

A system for smoke that's simulated in worldspace, so that it trails as the particle system moves

A system for embers - 20 particles max - these are smaller sparks coming out of the chimney! they have a little bit of orbit and a little bit of radial velocity, just to give 'em a little spice

- Fire from the chimney - 20 particles max - for when you're REALLY going fast! It's not a ton of fire, but it adds a nice soft glow to the chimney

I think that just about does it for this one - OH!

hang on

There are a couple of texture assets in the FBX folder - they spawned in there when I imported my FBXs, and I'm not sure if they NEED to be in there or if they could be moved to the materials folder. I figured better safe than sorry, and have elected to leave them there for the time being - sorry if this causes any confusion!

Anyhow, it's 9:37 am
So I'm going ot get some sleep here

keep up the good work, fellas!